requirements.txt

mlflow>=3.1
python-dotenv

tiktoken
langchain
langchain-openai
langchain-community

langchain-pinecone
pinecone
sentence-transformers